requirements.txt
fastapi 
uvicorn[standard]
python-dotenv 
openai  
aiohttp
sqlalchemy
psycopg2-binary
alembic